refactor(platform): declare the org notification link union once
The org notification `link` shape was written out in three places: the
router that turns a stored link into a route, the wire contract the app
consumes, and the producer, which typed it as `Record<string, unknown>`
and so accepted any shape at all.

Move it to `backend/core/notifications/org_notification_link.ts` and
import it in all three. The producer can no longer invent a `kind` the
router has never heard of.

Tightening the producer surfaced one existing widening:
`index-health.ts` built `link` inside a shared object literal, so `kind`
inferred as `string` rather than the literal. Pinned with `as const`,
matching the `category` field beside it.
